Use st_mtime instead of st_ctime
stat.st_ctime returns the creation time of the file on Windows (which doesn't change with new messages), and the modification time of the file *metadata* on Unix/Mac. stat.st_mtime should return the modification time of the file contents (i.e. time of last write to the file) across all platforms.